Solve exercise no.6 of the 'css-exercises' repo
Edit only the CSS file and make use of specificity and/or rule order
to achieve the required results. Add, remove, or edit selectors.

When conflicts arise, the term "cascading" refers to the hierarchical
order in which different style sheet types interact.

See 'README.md' file for more information about the solution path.